requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
langchain-chroma
langchain-experimental
langchain-huggingface
chromadb
feedparser
newspaper3k
apscheduler
python-dotenv
beautifulsoup4
lxml
sentence-transformers
